% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{get_logo}
\alias{get_logo}
\title{Download any Logo (url or local) to current working directory}
\usage{
get_logo(path)
}
\arguments{
\item{path}{File path or web links to an image file

 @export}
}
\description{
This function is a generic function that will download any image
 files to current working directory.
}
