% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memor.R
\name{memor}
\alias{memor}
\title{Curator Template}
\usage{
memor(use_profile = FALSE, logo = NULL, company = NULL,
  short_title = NULL, footer_on_first_page = TRUE, confidential = TRUE,
  toc = FALSE, lot = FALSE, lof = FALSE, number_sections = TRUE,
  latex_engine = "xelatex", ...)
}
\arguments{
\item{toc}{logical value for table of contents. See ?rmarkdown::pdf_document
for details}

\item{number_sections}{logical value for wether sections should be numbered.
See ?rmarkdown::pdf_document for details}

\item{latex_engine}{LaTeX engine. See ?rmarkdown::pdf_document for details}

\item{...}{other options to be passed to rmarkdown::pdf_document. See
?rmarkdown::pdf_document for details}
}
\description{
Curator Template
}
