% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memor_profile.R
\name{memor_profile}
\alias{memor_profile}
\title{Manage memor profile}
\usage{
memor_profile()
}
\description{
This RStudio Addin opens up the ~/memor-profile.yaml file for
users to edit. If the file doesn't exists, it will create one.
}
